+++ packages/liveManager/src/manager.ts
@@ -23,7 +23,7 @@ import {
   downloadImage,
   isBetweenTimeRange,
 } from "./utils.js";
-import { StreamManager } from "./streamManager.js";
+import { StreamManager } from "./recorder/streamManager.js";
 
 export interface RecorderProvider<E extends AnyObject> {
   // Provider 的唯一 id，最好只由英文 + 数字组成
@@ -75,21 +75,21 @@ export interface RecorderManager<
   E extends AnyObject = ME & PE,
 > extends Emitter<{
     error: { source: string; err: unknown };
-    RecordStart: { recorder: Recorder<E>; recordHandle: RecordHandle };
-    RecordSegment: { recorder: Recorder<E>; recordHandle?: RecordHandle };
-    videoFileCreated: { recorder: Recorder<E>; filename: string; cover?: string };
-    videoFileCompleted: { recorder: Recorder<E>; filename: string };
-    RecorderProgress: { recorder: Recorder<E>; progress: Progress };
+    RecordStart: { recorder: SerializedRecorder<E>; recordHandle: RecordHandle };
+    RecordSegment: { recorder: SerializedRecorder<E>; recordHandle?: RecordHandle };
+    videoFileCreated: { recorder: SerializedRecorder<E>; filename: string; cover?: string };
+    videoFileCompleted: { recorder: SerializedRecorder<E>; filename: string };
+    RecorderProgress: { recorder: SerializedRecorder<E>; progress: Progress };
     RecoderLiveStart: { recorder: Recorder<E> };
 
-    RecordStop: { recorder: Recorder<E>; recordHandle: RecordHandle; reason?: string };
-    Message: { recorder: Recorder<E>; message: Message };
+    RecordStop: { recorder: SerializedRecorder<E>; recordHandle: RecordHandle; reason?: string };
+    Message: { recorder: SerializedRecorder<E>; message: Message };
     RecorderUpdated: {
-      recorder: Recorder<E>;
+      recorder: SerializedRecorder<E>;
       keys: (string | keyof Recorder<E>)[];
     };
-    RecorderAdded: Recorder<E>;
-    RecorderRemoved: Recorder<E>;
+    RecorderAdded: SerializedRecorder<E>;
+    RecorderRemoved: SerializedRecorder<E>;
     RecorderDebugLog: DebugLog & { recorder: Recorder<E> };
     Updated: ConfigurableProp[];
   }> {
@@ -244,23 +244,32 @@ export function createRecorderManager<
       this.recorders.push(recorder);
 
       recorder.on("RecordStart", (recordHandle) =>
-        this.emit("RecordStart", { recorder, recordHandle }),
+        this.emit("RecordStart", { recorder: recorder.toJSON(), recordHandle }),
       );
       recorder.on("RecordSegment", (recordHandle) =>
-        this.emit("RecordSegment", { recorder, recordHandle }),
+        this.emit("RecordSegment", { recorder: recorder.toJSON(), recordHandle }),
       );
       recorder.on("videoFileCreated", ({ filename, cover }) => {
         if (recorder.saveCover && recorder?.liveInfo?.cover) {
           const coverPath = replaceExtName(filename, ".jpg");
           downloadImage(cover ?? recorder?.liveInfo?.cover, coverPath);
         }
-        this.emit("videoFileCreated", { recorder, filename });
+        this.emit("videoFileCreated", { recorder: recorder.toJSON(), filename });
       });
       recorder.on("videoFileCompleted", ({ filename }) =>
-        this.emit("videoFileCompleted", { recorder, filename }),
+        this.emit("videoFileCompleted", { recorder: recorder.toJSON(), filename }),
+      );
+      recorder.on("Message", (message) =>
+        this.emit("Message", { recorder: recorder.toJSON(), message }),
+      );
+      recorder.on("Updated", (keys) =>
+        this.emit("RecorderUpdated", { recorder: recorder.toJSON(), keys }),
+      );
+      recorder.on("DebugLog", (log) =>
+        this.emit("RecorderDebugLog", { recorder: recorder, ...log }),
       );
       recorder.on("RecordStop", ({ recordHandle, reason }) => {
-        this.emit("RecordStop", { recorder, recordHandle, reason });
+        this.emit("RecordStop", { recorder: recorder.toJSON(), recordHandle, reason });
         // 如果reason中存在"invalid stream"，说明直播由于某些原因中断了，虽然会在下一次周期检查中继续，但是会遗漏一段时间。
         // 这时候可以触发一次检查，但出于直播可能抽风的原因，为避免风控，一场直播最多触发五次。
         // 测试阶段，还需要一个开关，默认关闭，几个版本后转正使用
@@ -297,11 +306,8 @@ export function createRecorderManager<
           }, 1000);
         }
       });
-      recorder.on("Message", (message) => this.emit("Message", { recorder, message }));
-      recorder.on("Updated", (keys) => this.emit("RecorderUpdated", { recorder, keys }));
-      recorder.on("DebugLog", (log) => this.emit("RecorderDebugLog", { recorder, ...log }));
       recorder.on("progress", (progress) => {
-        this.emit("RecorderProgress", { recorder, progress });
+        this.emit("RecorderProgress", { recorder: recorder.toJSON(), progress });
       });
       recorder.on("videoFileCreated", () => {
         if (!recorder.liveInfo?.liveId) return;
@@ -310,10 +316,10 @@ export function createRecorderManager<
         if (liveStartObj[key]) return;
         liveStartObj[key] = true;
 
-        this.emit("RecoderLiveStart", { recorder });
+        this.emit("RecoderLiveStart", { recorder: recorder });
       });
 
-      this.emit("RecorderAdded", recorder);
+      this.emit("RecorderAdded", recorder.toJSON());
 
       return recorder;
     },
@@ -324,7 +330,7 @@ export function createRecorderManager<
       this.recorders.splice(idx, 1);
 
       delete tempBanObj[recorder.channelId];
-      this.emit("RecorderRemoved", recorder);
+      this.emit("RecorderRemoved", recorder.toJSON());
     },
     async startRecord(id: string) {
       const recorder = this.recorders.find((item) => item.id === id);