+++ packages/liveManager/src/recorder.ts
@@ -56,6 +56,8 @@ export interface RecorderCreateOpts<E extends AnyObject = UnknownObject> {
   titleKeywords?: string;
   /** 用于指定录制文件格式，auto时，分段使用ts，不分段使用mp4 */
   videoFormat?: "auto" | "ts" | "mkv";
+  /** 录制类型 */
+  recorderType?: "ffmpeg" | "mesio";
   /** 流格式优先级 */
   formatriorities?: Array<"flv" | "hls">;
   /** 只录制音频 */
@@ -68,7 +70,27 @@ export interface RecorderCreateOpts<E extends AnyObject = UnknownObject> {
   extra?: Partial<E>;
 }
 
-export type SerializedRecorder<E extends AnyObject> = PickRequired<RecorderCreateOpts<E>, "id">;
+export type SerializedRecorder<E extends AnyObject> = PickRequired<RecorderCreateOpts<E>, "id"> &
+  Pick<
+    Recorder<E>,
+    | "id"
+    | "channelId"
+    | "remarks"
+    | "disableAutoCheck"
+    | "quality"
+    | "streamPriorities"
+    | "sourcePriorities"
+    | "extra"
+    | "segment"
+    | "saveSCDanma"
+    | "saveCover"
+    | "saveGiftDanma"
+    | "disableProvideCommentsWhenRecording"
+    | "liveInfo"
+    | "uid"
+    | "titleKeywords"
+    // | "recordHandle"
+  >;
 
 export type RecorderState = "idle" | "recording" | "stopping-record";
 export type Progress = { time: string | null };