+++ packages/DouYinRecorder/CHANGELOG.md
@@ -1,5 +1,7 @@
-# Next
+# 1.6.0
 
+- 支持 `recorderType` 参数用于配置底层录制器，支持`ffmpeg | mesio`
+- 修复 `videoFormat=auto` 时某些情况下格式的判断
 - 支持用户主页解析
 
 # 1.5.3