+++ packages/HuYaRecorder/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.6.0
+
+- 支持 `recorderType` 参数用于配置底层录制器，支持`ffmpeg | mesio`
+- 修复 `videoFormat=auto` 时某些情况下格式的判断
+
 # 1.3.2
 
 - 修复星秀区录制原画可能失败的bug